requirements.txt
# Web Framework & Utilities
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]==2.12.5
python-multipart
python-dotenv
PyYAML

# DS ML
torch
numpy
pandas
scikit-learn
scipy
sentence-transformers>=2.3.1

# Vector Database
faiss-cpu

# Database
pymongo[srv]

# Testing
pytest
pytest-asyncio
httpx
mongomock==4.1.2  # mocking MongoDB 
pytest-mock==3.11.1
# unittest included in standard lib


# Auth
python-jose
python-jose[cryptography]
passlib       # hasing algo